Server Updates + Reorganization
Moved all of the icons sitting in the home directory to a icon folder (except for the favicon).

Also, the server got a favicon and logo!

The logo can be seen on the welcome page slideshow, in place of the Virtual Machine image.
Favicons should be on all pages now.

Unfortunately, it was decided to end the Virtual Machine service. It wasn't all that useful, seeing as the machines hosted on copy.sh were dating back to the 90's!  You couldn't really test your programs on them.