Added new feature to welcome new users
Updates were made to the registration and login pages, to determine if a user is new. If they are, they will get a little "greeting", a special welcome page that only displays if they are new.

If the user isn't new to MyPortal, they will be given the normal welcome page.