Fixes to registration and minor visual changes
Fixed the registration once and for all! It now uses a SESSION variable to determine if the user is new (it won't work if the user closes their browser then logs back in).

Added MyPortal logo to bottom of login, registration, and the new user welcome page.

Just hit version 1.2.0 !!!